Safari-611.1.10/LayoutTests/fast/css/MediaList-mediaText-null-expected.txt

Tests that MediaList.mediaText is not nullable and treats null as the empty string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaList.__proto__ is MediaList.prototype
PASS mediaList.length is 1
PASS mediaList.mediaText is "screen and (min-width: 480px)"
PASS mediaList.mediaText = null did not throw exception.
PASS mediaList.mediaText is ""
PASS mediaList.mediaText = 'screen and (min-width: 480px)' did not throw exception.
PASS mediaList.mediaText is "screen and (min-width: 480px)"
PASS style.removeAttribute('media') did not throw exception.
PASS mediaList.mediaText is ""
PASS successfullyParsed is true

TEST COMPLETE
